Enable second motor IC + fix logic stupidity
First, write HIGH to all motor enable pins. They were previously removed
because I though they were unneeded, but they actually are.

Second, enable the second motor IC so that we can control up and down.

Third, fix the weird ternary if in the calls to setMotor. When evaluating
when to reverse the motor, we compare against 255, which is always true unless
it's max. Set it to less than 0, which is the correct value.